Implement a score display view.
Scores can be displayed after the whole experiment in a separate view.
Each test can store a general score that is included in the totalscore,
and scores from each phase are also listed.

It is also possible to store a named score e.g "number of correct answers"
which will be displayed in the score list below the main score.