Implement batching of mongo queries
The mongo persistor will by default only return a certain number
of collections per message (configured by mongos batch_size). If
the batch_size is exeeded we'll sen a reply to the persistor with a
new respons handler.

This enables us to fetch as much data as we want in a single query,
as long as we dont run out of memory.