Improve component loading and added us locale
Should fix issue #9

Dojo seems to try to load the user's locale, which somehow resulted in
dojo trying to load 2 different locales in IE9. Which led to some missing
value... Setting the locale to a fixed on fixes this, but we loose i18n at
the same time.

The issue may be caused by the custom components, since they might not
be able to utilize dojos builtin localization features.