Add support for test folders.
A folder can be specified for all tests, which is saved on change.
Folders are available in the testlist view, but these should also be
ported to the main view.

Changed how the test name is saved as well to unify everything about
test saving. Thus avoiding alot of repetition.