Implement wait on refresh when showing something
Calling show(stuff pos true) will issue a wait command that will
be resumed on the next requestAnimationFrame. This should improve
the timing precision when showing time sensitive stimuli. However
browser support seems to be a bit limited at the moment, but IE
Edge handles this perfectly with a standard deviation of 0.5ms.
Chrome and FF benefit a little in form of better standarddeviation
but the sync doesn't seem to be tied completely to vsync. Well chrome
does something but not quite sure how that works.