Implement phase aggregation functionality
This function counts how many submitted dataobjects there are for
each phase, and aggregates them, Requires the latest version of
the mongopersistor which is not available in the vertx mod repository.
So this should be installed manually, or uploaded to a repo.

Anyways, old functionality will work just fine without the update.