Implement drag and drop functionality in tests
Possible to mark objects as draggable, and give them data that is passed
when dropped in a dropzone.

Dropzones define a function to call when something is dropped in them, and
the data specified on a draggable obejct is passed to this function as
well. Dropzones will also light up when a draggable object is dragged
above it.