Fix test progression when logged in + move stuff around
Test progression should work properly everywhere when logged in. And the
right id, wether logged in or not should always be used.

Also move the end and the json functions to the new db syntax. The only
thing remaining in the old file are the data export functions, which will
be updated and improved at some point anyways