Add function to manually enroll users in trainings
The testleader can manually enroll users in training experiments. This is
done by entering the users ID(s), comma separated, at the admin screen.
This is equivalent to clicking participate as a user. The training
experiment will be visible after the user has logged in.

Note that this is only doable with registered users.