Fix db session auth so that it works with postrequest
Post request can have a data hadnler callback to handle postdata. This handler should be called immeditially when processing the request, otherwise no data will be recieved. 

Fixed this by calling the routehandler when the requests is recieved, which enables the datahandler, and then waiting for the session check to occure, before calling the endfunction. This is only done if the user has a sessioncookie and dosn't have a session stored in memory.