Implement reminder emails for training participants
The testleader can enable reminder emails in the training settings. Reminder emails are sent to the training participants x hours after the last completed training round. 

We'll only send one reminder email per round for now, but that can be changed if needed...

The mailing function is run every 12 hours.