Add support for randomized trainings
This allows the testleader to specify a random order for the training and control phases. It also works in tandem with iterationcontrol. The random order is generated when a user is added to the training experiment (particiaptes), so it's is strongly recommended that the component order is not changed in an ongoing training experiment. We only store the random order for the relevant phases, so the random order array can be a bit hard to read. This could  however maybe be fixed in the data export layer.