Create required experiment folders when experiment is created
The lazy way where we'de create the folder on the fly when streaming the upload didn't work. Having
an async callback before callling upload.streamtofilesystem resulted in missing bytes at the beginning
of the file. The cause for this is that the filestream is initiated in the upload handler, but we start
piping the stream to the fs after in the callback, shich means that there is nothing that catches
the beginning of the stream. So that just dissapears. There's probably a proper workaround for this
if you use read/write-streams correctly, but this solution makes things less complicated.